numaflow-dra/dci_poc/pyproject.toml
[tool.poetry]
name = "dci-poc"
version = "1.0.3"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pynumaflow = "0.10.0"

python-dotenv = "1.0.1"
opencv-python = "4.10.0.84"
numpy = "2.2.2"
PyYAML = "6.0.2"
pathlib = "1.0.1"
torch = "2.6.0"
PyTurboJPEG = "1.7.7"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
pytest-cov = "^3.0"
pytest-mock = "^3.14.1"
